refactor(app): add `utils` folder, update paths
