+++ src/main.ts
@@ -3,7 +3,7 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
-import { Logger } from './app/infrastructure/core/services/logger';
+import { Logger } from './app/infrastructure/utils/logger';
 
 if (environment.production) {
   enableProdMode();
